closes #1199
